==> Data_Analysis/Analisis_de_datos/ch_04/2-operaciones_con_dataframe.ipynb <==
# Operaciones DataFrame

## Acerca de los datos
En este cuaderno trabajaremos con 2 conjuntos de datos:
- La cotización de las acciones de Facebook a lo largo de 2018 (obtenida mediante la [`stock_analysis` package](https://github.com/stefmolin/stock-analysis)).
- Datos meteorológicos diarios de Nueva York[National Centers for Environmental Information (NCEI) API](https://www.ncdc.noaa.gov/cdo-web/webservices/v2).

*Nota: El NCEI forma parte de la Administración Nacional Oceánica y Atmosférica (NOAA) y, como puede ver en la URL de la API, este recurso se creó cuando el NCEI se llamaba NCDC. Si la URL de este recurso cambiara en el futuro, puede buscar "NCEI weather API" para encontrar la actualizada.*

## Antecedentes de los datos meteorológicos

Significado de los datos:
- `AWND`: velocidad media del viento
- `PRCP`: precipitación en milímetros
- `SNOW`: nevadas en milímetros
- SNWD`: profundidad de la nieve en milímetros
- TMAX`: temperatura máxima diaria en grados Celsius
- TMIN`: temperatura mínima diaria en grados Celsius

## Setup

import numpy as np
import pandas as pd

weather = pd.read_csv('data/nyc_weather_2018.csv', parse_dates=['date'])
weather.head()

fb = pd.read_csv('data/fb_2018.csv', index_col='date', parse_dates=True)
fb.head()

## Aritmética y estadística
Ya vimos que podemos utilizar operadores matemáticos como `+` y `/` con dataframes directamente. Sin embargo, también podemos utilizar métodos, que nos permiten especificar el eje sobre el que realizar el cálculo. Por defecto, esto es por columna. Busquemos las puntuaciones Z para el volumen negociado y veamos los días en los que éste se alejó más de 3 desviaciones estándar de la media:

fb.assign(
    abs_z_score_volume=lambda x: \
        x.volume.sub(x.volume.mean()).div(x.volume.std()).abs()
).query('abs_z_score_volume > 3')

Podemos utilizar `rank()` y `pct_change()` para ver qué días tuvieron el mayor cambio en el volumen negociado desde el día anterior:

fb.assign(
    volume_pct_change=fb.volume.pct_change(),
    pct_change_rank=lambda x: \
        x.volume_pct_change.abs().rank(ascending=False)
).nsmallest(5, 'pct_change_rank')

El 12 de enero se conoció la noticia de que Facebook cambiaba su feed de noticias para centrarse más en los contenidos de los amigos de los usuarios que en las marcas que siguen. Dado que la publicidad de Facebook es un componente clave de su negocio ([nearly 89% in 2017](https://www.investopedia.com/ask/answers/120114/how-does-facebook-fb-make-money.asp)) se vendieron muchas acciones y la cotización cayó presa del pánico:

fb['2018-01-11':'2018-01-12']

A lo largo de 2018, el precio de las acciones de Facebook nunca tuvo un mínimo por encima de 215 dólares:

(fb > 215).any()

Todos los precios OHLC (apertura, máximo, mínimo y cierre) de Facebook estuvieron al menos un día a 215 dólares o menos:

(fb > 215).all()

## Binning
Cuando trabajamos con volumen negociado, puede interesarnos más el rango de volumen que los valores exactos. No hay dos días con el mismo volumen negociado:

(fb.volume.value_counts() > 1).sum()

Podemos utilizar `pd.cut()` para crear 3 bins de rango par en volumen negociado y nombrarlos. A continuación, podemos trabajar con categorías de volumen negociado bajo, medio y alto:

volume_binned = pd.cut(fb.volume, bins=3, labels=['low', 'med', 'high'])
volume_binned.value_counts()

Veamos los días con mayor volumen de negociación:

fb[volume_binned == 'high'].sort_values('volume', ascending=False)

El 25 de julio, Facebook anunció un crecimiento decepcionante del número de usuarios y las acciones se desplomaron a última hora:

fb['2018-07-25':'2018-07-26']

El escándalo de Cambridge Analytica estalló el sábado 17 de marzo, así que nos fijamos en el lunes posterior para conocer las cifras:

fb['2018-03-16':'2018-03-20']

Como la mayoría de los días tienen un volumen similar, pero unos pocos son muy grandes, tenemos intervalos muy amplios. La mayoría de los datos se encuentran en la zona baja.

from visual-aids.visual_aids.misc_viz import low_med_high_bins_viz

low_med_high_bins_viz(
    fb, 'volume', ylabel='volume traded',
    title='Daily Volume Traded of Facebook Stock in 2018 (with bins)'
)

Si dividimos utilizando cuantiles, los intervalos tendrán aproximadamente el mismo número de observaciones. Para ello, utilizaremos `qcut()`. Haremos 4 cuartiles:

volume_qbinned = pd.qcut(fb.volume, q=4, labels=['q1', 'q2', 'q3', 'q4'])
volume_qbinned.value_counts()

Fíjate en que los contenedores ya no cubren rangos del mismo tamaño:

from visual-aids.visual_aids.misc_viz import quartile_bins_viz

quartile_bins_viz(
    fb, 'volume', ylabel='volume traded', 
    title='Daily Volume Traded of Facebook Stock in 2018 (with quartile bins)'
)

## Aplicar Funciones
Podemos utilizar el método `apply()` para ejecutar la misma operación en todas las columnas (o filas) del marco de datos. En primer lugar, vamos a aislar las observaciones meteorológicas de la estación de Central Park y a pivotar los datos:

central_park_weather = weather\
    .query('station == "GHCND:USW00094728"')\
    .pivot(index='date', columns='datatype', values='value')

Calculemos las puntuaciones Z de las observaciones TMIN, TMAX y PRCP en Central Park en octubre de 2018:

oct_weather_z_scores = central_park_weather\
    .loc['2018-10', ['TMIN', 'TMAX', 'PRCP']]\
    .apply(lambda x: x.sub(x.mean()).div(x.std()))
oct_weather_z_scores.describe().T

El 27 de octubre llovió mucho más que el resto de los días:

oct_weather_z_scores.query('PRCP > 3').PRCP

De hecho, este día fue mucho más alto que el resto:

central_park_weather.loc['2018-10', 'PRCP'].describe()

Cuando la función que queremos aplicar no está vectorizada, podemos
- utilizar `np.vectorize()` para vectorizarla (de forma similar a como funciona `map()`) y luego utilizarla con `apply()`.
- utilizar `applymap()` y pasarle directamente la función no vectorizada

Digamos que queremos contar los dígitos de los números enteros de los datos de Facebook; `len()` no está vectorizada, así que podemos usar `np.vectorize()` o `applymap()`:

fb.apply(
    lambda x: np.vectorize(lambda y: len(str(np.ceil(y))))(x)
).astype('int64').equals(
    fb.applymap(lambda x: len(str(np.ceil(x))))
)

Una simple operación de suma a cada elemento de una serie crece linealmente en complejidad temporal cuando se utiliza `iteritems()`, pero se mantiene cerca de 0 cuando se utilizan operaciones vectorizadas. `iteritems()` y los métodos relacionados sólo deben utilizarse si no existe una solución vectorizada:

import time

import numpy as np
import pandas as pd

np.random.seed(0)

vectorized_results = {}
iteritems_results = {}

for size in [10, 100, 1000, 10000, 100000, 500000, 1000000, 5000000, 10000000]:
    # set of numbers to use
    test = pd.Series(np.random.uniform(size=size))
    
    # time the vectorized operation
    start = time.time()
    x = test + 10
    end = time.time()
    vectorized_results[size] = end - start
    
    # time the operation with `iteritems()`
    start = time.time()
    x = []
    for i, v in test.iteritems():
        x.append(v + 10)
    x = pd.Series(x)
    end = time.time()
    iteritems_results[size] = end - start

results = pd.DataFrame(
    [pd.Series(vectorized_results, name='vectorized'), pd.Series(iteritems_results, name='iteritems')]
).T    

# plotting
ax = results.plot(title='Time Complexity', color=['blue', 'red'], legend=False)

# formatting
ax.set(xlabel='item size (rows)', ylabel='time (s)')
ax.text(0.5e7, iteritems_results[0.5e7] * .9, 'iteritems()', rotation=34, color='red', fontsize=12, ha='center', va='bottom')
ax.text(0.5e7, vectorized_results[0.5e7], 'vectorized', color='blue', fontsize=12, ha='center', va='bottom')
for spine in ['top', 'right']:
    ax.spines[spine].set_visible(False)

## Window Calculations

*Consultar el [`understanding_window_calculations.ipynb`](./understanding_window_calculations.ipynb) para visualizaciones interactivas mediante widgets que ayudan a comprender los cálculos de las ventanas.*

El método `rolling()` nos permite realizar cálculos de ventanas móviles. Simplemente especificamos el tamaño de la ventana (3 días aquí) y lo seguimos con una llamada a una función de agregación (suma aquí):

central_park_weather.loc['2018-10'].assign(
    rolling_PRCP=lambda x: x.PRCP.rolling('3D').sum()
)[['PRCP', 'rolling_PRCP']].head(7).T

También podemos realizar los cálculos continuos en todo el marco de datos a la vez. Esto aplicará la misma función de agregación a cada columna:

central_park_weather.loc['2018-10'].rolling('3D').mean().head(7).iloc[:,:6]

Podemos utilizar diferentes funciones de agregación por columna si utilizamos `agg()` en su lugar. Pasamos un diccionario asignando la columna a la agregación a realizar sobre ella. Aquí, unimos el resultado a los datos originales para ver qué ocurre:

central_park_weather['2018-10-01':'2018-10-07'].rolling('3D').agg(
    {'TMAX': 'max', 'TMIN': 'min', 'AWND': 'mean', 'PRCP': 'sum'}
).join( # unir con los datos originales para comparar
    central_park_weather[['TMAX', 'TMIN', 'AWND', 'PRCP']], 
    lsuffix='_rolling'
).sort_index(axis=1) # ordena las columnas para que los calcos rodantes estén junto a los originales

Supongamos que reindexamos los datos de las acciones de Facebook como hicimos con los datos del S&P 500 en el capítulo 3. Si utilizáramos cálculos continuos con estos datos, incluiríamos los valores cuando el mercado estuviera cerrado:

fb_reindexed = fb\
    .reindex(pd.date_range('2018-01-01', '2018-12-31', freq='D'))\
    .assign(
        volume=lambda x: x.volume.fillna(0),
        close=lambda x: x.close.fillna(method='ffill'),
        open=lambda x: x.open.combine_first(x.close),
        high=lambda x: x.high.combine_first(x.close),
        low=lambda x: x.low.combine_first(x.close)
    )
fb_reindexed.assign(day=lambda x: x.index.day_name()).head(10)

A partir de la versión 1.0, `pandas` soporta la definición de ventanas personalizadas para cálculos rolling, lo que nos permite realizar cálculos rolling en los días en que el mercado estuvo abierto. Una forma es crear una nueva clase que herede de `BaseIndexer` y proporcionar la lógica para determinar los límites de la ventana en la clase `get_window_bounds()` (más información [here](https://pandas.pydata.org/pandas-docs/stable/user_guide/computation.html#custom-window-rolling)). En nuestro caso, podemos utilizar la clase `VariableOffsetWindowIndexer`, que se introdujo en la versión 1.1, para realizar cálculos continuos sobre desfases temporales no fijos (como días laborables). Realicemos un cálculo móvil de tres días hábiles sobre los datos de acciones de Facebook reindexados y unámoslos a los datos reindexados para compararlos:

from pandas.api.indexers import VariableOffsetWindowIndexer

indexer = VariableOffsetWindowIndexer(
    index=fb_reindexed.index, offset=pd.offsets.BDay(3)
)
fb_reindexed.assign(window_start_day=0).rolling(indexer).agg({
    'window_start_day': lambda x: x.index.min().timestamp(),
    'open': 'mean', 'high': 'max', 'low': 'min',
    'close': 'mean', 'volume': 'sum'
}).join(
    fb_reindexed, lsuffix='_rolling'
).sort_index(axis=1).assign(
    day=lambda x: x.index.day_name(),
    window_start_day=lambda x: pd.to_datetime(x.window_start_day, unit='s')
).head(10)

Los cálculos continuos `rolling()` utilizan una ventana deslizante. Los cálculos expansivos `expanding()`, sin embargo, crecen en tamaño. Son equivalentes a agregaciones acumulativas como `cumsum()`; sin embargo, podemos especificar el número mínimo de periodos necesarios para empezar a calcular (por defecto es 1), y no estamos limitados a agregaciones predefinidas. Por lo tanto, aunque no existe un método para la media acumulada, podemos calcularla utilizando `expanding()`. Calculemos la precipiación media del mes hasta la fecha:

central_park_weather.loc['2018-06'].assign(
    TOTAL_PRCP=lambda x: x.PRCP.cumsum(),
    AVG_PRCP=lambda x: x.PRCP.expanding().mean()
).head(10)[['PRCP', 'TOTAL_PRCP', 'AVG_PRCP']].T

También podemos usar `agg()` para especificar agregaciones por columna. Tenga en cuenta que esto también funciona con las funciones NumPy. Aquí, unimos los cálculos de expansión con los resultados originales para compararlos:

central_park_weather['2018-10-01':'2018-10-07'].expanding().agg(
    {'TMAX': np.max, 'TMIN': np.min, 'AWND': np.mean, 'PRCP': np.sum}
).join(
    central_park_weather[['TMAX', 'TMIN', 'AWND', 'PRCP']], 
    lsuffix='_expanding'
).sort_index(axis=1)

Pandas proporciona el método `ewm()` para cálculos de medias móviles ponderadas exponencialmente. Como vimos en el capítulo 1, podemos utilizar la media móvil ponderada exponencialmente para suavizar los datos. Comparemos la media móvil con la media móvil ponderada exponencialmente con la temperatura máxima diaria. Nótese que `span` aquí son los periodos a utilizar:

central_park_weather.assign(
    AVG=lambda x: x.TMAX.rolling('30D').mean(),
    EWMA=lambda x: x.TMAX.ewm(span=30).mean()
).loc['2018-09-29':'2018-10-08', ['TMAX', 'EWMA', 'AVG']].T

*Consultar el [`understanding_window_calculations.ipynb`](./understanding_window_calculations.ipynb) para visualizaciones interactivas que ayudan a comprender los cálculos de las ventanas.*

## Pipes
Las tuberías son una forma de racionalizar nuestro código `pandas` y hacerlo más legible y flexible. Usando tuberías, podemos tomar una llamada anidada como

```python
f(g(h(data), 20), x=True)
```

y convertirlo en algo más legible:

```python
data.pipe(h)\
    .pipe(g, 20)\
    .pipe(f, x=True)\
```

Podemos utilizar tuberías para aplicar cualquier función que acepte nuestros datos como primer argumento y pasar cualquier argumento adicional. Esto facilita encadenar pasos independientemente de si son métodos o funciones:

Podemos pasar cualquier función que acepte la llamada de `pipe()` como primer argumento:

def get_info(df):
    return '%d rows, %d columns and max closing Z-score was %d' % (*df.shape, df.close.max())

get_info(fb.loc['2018-Q1'].apply(lambda x: (x - x.mean())/x.std()))\
    == fb.loc['2018-Q1'].apply(lambda x: (x - x.mean())/x.std()).pipe(get_info)

Por ejemplo, pasar `pd.DataFrame.rolling` a `pipe()` es equivalente a llamar a `rolling()` directamente en el dataframe, excepto que tenemos más flexibilidad para cambiar esto:

fb.pipe(pd.DataFrame.rolling, '20D').mean().equals(fb.rolling('20D').mean())

La tubería toma la función pasada y la llama con el objeto que llamó a `pipe()` como primer argumento. Los argumentos posicionales y de palabra clave se pasan hacia abajo:

pd.DataFrame.rolling(fb, '20D').mean().equals(fb.rolling('20D').mean())

Podemos utilizar una tubería para crear una función que podamos utilizar para todas nuestras necesidades de cálculo de ventanas:

from window_calc import window_calc
window_calc??

Ahora podemos utilizar la misma interfaz para realizar diversos cálculos de ventanas. Busquemos la mediana en expansión de los datos de Facebook:

window_calc(fb, pd.DataFrame.expanding, np.median).head()

El uso de la media móvil ponderada exponencialmente requiere que pasemos un argumento de palabra clave:

window_calc(fb, pd.DataFrame.ewm, 'mean', span=3).head()

Con los cálculos rodantes, podemos pasar un argumento posicional para el tamaño de la ventana:

window_calc(
    central_park_weather.loc['2018-10'], 
    pd.DataFrame.rolling, 
    {'TMAX': 'max', 'TMIN': 'min', 'AWND': 'mean', 'PRCP': 'sum'},
    '3D'
).head()


<div style="display: flex; justify-content: space-between; margin-bottom: 10px;">
    <div style="text-align: left;">
        <a href="./1-consulta_y_merge.ipynb">
            <button>&#8592; Notebook Anterior</button>
        </a>
    </div>
    <div style="text-align: center;">
        <a href="./calculo_de_ventana.ipynb">
            <button>Calculo de Ventana</button>
        </a>
    </div>
    <div style="text-align: right;">
        <a href="./3-agregaciones.ipynb">
            <button>Proximo Notebook &#8594;</button>
        </a>
    </div>
</div>

<hr>
