==> Data_Analysis/Analisis_de_datos/ch_04/4-time_series.ipynb <==
# Trabajar con datos de series temporales

## Acerca de los datos
En este cuaderno trabajaremos con 5 conjuntos de datos:
- (CSV) Cotización diaria de las acciones de Facebook a lo largo de 2018 (obtenida mediante el método [`stock_analysis` package](https://github.com/stefmolin/stock-analysis)).
- (CSV) Datos bursátiles OHLC de Facebook del 20 de mayo de 2019 al 24 de mayo de 2019 por minuto de [Nasdaq.com](https://old.nasdaq.com/symbol/fb/interactive-chart).
- (CSV) datos de acciones fundidos para Facebook del 20 de mayo de 2019 al 24 de mayo de 2019 por minuto de [Nasdaq.com](https://old.nasdaq.com/symbol/fb/interactive-chart).
- (DB) precios de apertura de las acciones por minuto para Apple del 20 de mayo de 2019 al 24 de mayo de 2019 alterados para tener segundos en el tiempo de [Nasdaq.com](https://old.nasdaq.com/symbol/aapl/interactive-chart).
- (DB) precios de apertura de las acciones de Facebook por minuto del 20 de mayo de 2019 al 24 de mayo de 2019 de [Nasdaq.com](https://old.nasdaq.com/symbol/fb/interactive-chart).

## Setup

import numpy as np
import pandas as pd

fb = pd.read_csv('data/fb_2018.csv', index_col='date', parse_dates=True).assign(
    trading_volume=lambda x: pd.cut(x.volume, bins=3, labels=['low', 'med', 'high'])
)
fb.head()

## Selección y filtrado por tiempo
Recuerda, cuando tenemos un índice de tipo `DatetimeIndex`, podemos utilizar el corte por fechas. Podemos proporcionar un rango de fechas. Sólo recuperamos tres días porque la bolsa cierra los fines de semana:

fb['2018-10-11':'2018-10-15']

Podemos seleccionar rangos de meses y trimestres:

fb.loc['2018-q1'].equals(fb['2018-01':'2018-03'])

El método `first()` nos dará una longitud de tiempo especificada desde el principio de la serie temporal. Aquí, pedimos una semana. El 1 de enero de 2018 era festivo, es decir, el mercado estaba cerrado. También fue un lunes, por lo que la semana aquí es de sólo cuatro días:

fb.first('1W')

El método `last()` tomará del final:

fb.last('1W')

Supongamos que reindexamos los datos de las acciones de Facebook para incluir todas las fechas de 2018. Tendríamos entradas nulas para el 1 de enero:

fb_reindexed = fb.reindex(pd.date_range('2018-01-01', '2018-12-31', freq='D'))
fb_reindexed.first('1D').isna().squeeze().all()

Podemos utilizar `first_valid_index()` para obtener el índice de la primera entrada no nula de nuestros datos, que es el primer día de apertura del mercado en el primer trimestre de 2018:

fb_reindexed.loc['2018-Q1'].first_valid_index()

A la inversa, podemos utilizar `last_valid_index()` para obtener la última entrada de datos no nulos. Para el primer trimestre de 2018, es el 29 de marzo:

fb_reindexed.loc['2018-Q1'].last_valid_index()

Podemos utilizar `asof()` para encontrar el último dato no nulo anterior al punto que buscamos. Si pedimos el 31 de marzo, obtendremos los datos del índice que obtuvimos de `fb_reindexed.loc['2018-Q1'].last_valid_index()`, que fue el 29 de marzo. Tenga en cuenta que esto funciona independientemente de si hemos reindexado:

fb_reindexed.asof('2018-03-31')

Para los siguientes ejemplos, necesitamos fechas y horas, por lo que leeremos el archivo de datos de acciones por minuto:

stock_data_per_minute = pd.read_csv(
    'data/fb_week_of_may_20_per_minute.csv', index_col='date', parse_dates=True, 
    date_parser=lambda x: pd.to_datetime(x, format='%Y-%m-%d %H-%M')
)

stock_data_per_minute.head()

Podemos utilizar un objeto `Grouper` para enrollar nuestros datos a nivel diario junto con `first` y `last`:

stock_data_per_minute.groupby(pd.Grouper(freq='1D')).agg({
    'open': 'first',
    'high': 'max', 
    'low': 'min', 
    'close': 'last', 
    'volume': 'sum'
})

El método `at_time()` nos permite extraer todas las fechas que coincidan con una hora determinada. En este caso, podemos obtener todas las filas a partir de la hora de apertura de la bolsa (9:30 AM):

stock_data_per_minute.at_time('9:30')

Podemos utilizar `between_time()` para obtener los datos de los dos últimos minutos de la negociación diaria:

stock_data_per_minute.between_time('15:59', '16:00')

Por término medio, ¿se negocian más acciones en los primeros 30 minutos de negociación o en los últimos 30 minutos? Podemos combinar `between_time()` con `group_by()` y `filter()` de la función [`3-aggregations.ipynb`](./3-aggregations.ipynb) para responder a esta pregunta. En la semana en cuestión, se negocia más por término medio a la hora de apertura que a la hora de cierre:

shares_traded_in_first_30_min = stock_data_per_minute\
    .between_time('9:30', '10:00')\
    .groupby(pd.Grouper(freq='1D'))\
    .filter(lambda x: (x.volume > 0).all())\
    .volume.mean()

shares_traded_in_last_30_min = stock_data_per_minute\
    .between_time('15:30', '16:00')\
    .groupby(pd.Grouper(freq='1D'))\
    .filter(lambda x: (x.volume > 0).all())\
    .volume.mean()

shares_traded_in_first_30_min - shares_traded_in_last_30_min

En los casos en que la hora no importa, podemos normalizar las horas a medianoche:

pd.DataFrame(
    dict(before=stock_data_per_minute.index, after=stock_data_per_minute.index.normalize())
).head()

Tenga en cuenta que también podemos utilizar `normalize()` en un objeto `Series` después de acceder al atributo `dt`:

stock_data_per_minute.index.to_series().dt.normalize().head()

## Desplazamiento para datos retardados
Podemos utilizar `shift()` para crear datos retardados. Por defecto, el desplazamiento será de un periodo. Por ejemplo, podemos utilizar `shift()` para crear una nueva columna que indique el precio de cierre del día anterior. A partir de esta nueva columna, podemos calcular el cambio de precio debido a la negociación fuera de horario (después del cierre de un día hasta la apertura del día siguiente):

fb.assign(
    prior_close=lambda x: x.close.shift(),
    after_hours_change_in_price=lambda x: x.open - x.prior_close,
    abs_change=lambda x: x.after_hours_change_in_price.abs()
).nlargest(5, 'abs_change')

Si el objetivo es sumar/restar tiempo, podemos utilizar objetos `pd.Timedelta` en su lugar:

pd.date_range('2018-01-01', freq='D', periods=5) + pd.Timedelta('9 hours 30 minutes')

## Datos diferenciados
Utilizar el método `diff()` es una forma rápida de calcular la diferencia entre los datos y una versión retardada de los mismos. Por defecto, dará el resultado de `data - data.shift()`:

(
    fb.drop(columns='trading_volume') 
    - fb.drop(columns='trading_volume').shift()
).equals(
    fb.drop(columns='trading_volume').diff()
)

Podemos utilizarlo para ver la evolución diaria de las acciones de Facebook:

fb.drop(columns='trading_volume').diff().head()

Podemos especificar el número de períodos, puede ser cualquier número entero positivo o negativo:

fb.drop(columns='trading_volume').diff(-3).head()

## Remuestreo
A veces los datos tienen una granularidad que no es propicia para nuestro análisis. Consideremos el caso en el que tenemos datos por minuto para todo el año 2018. Veamos qué ocurre si intentamos graficar esto, y luego observamos la agregación diaria de estos datos.

from visual-aids.visual_aids.misc_viz import resampling_example
resampling_example()

El gráfico de la izquierda tiene tantos datos que no se ve nada. Sin embargo, cuando agregamos los totales diarios, vemos los datos. Podemos modificar la granularidad de los datos con los que trabajamos utilizando el remuestreo. Recordemos nuestros datos bursátiles minuto a minuto:

stock_data_per_minute.head()

Podemos remuestrear esto para llegar a una frecuencia diaria:

stock_data_per_minute.resample('1D').agg({
    'open': 'first',
    'high': 'max', 
    'low': 'min', 
    'close': 'last', 
    'volume': 'sum'
})

Podemos reducir la muestra a datos trimestrales:

fb.resample('Q').mean()

También podemos utilizar `apply()`. Aquí, mostramos el cambio trimestral de principio a fin:

fb.drop(columns='trading_volume').resample('Q').apply(
    lambda x: x.last('1D').values - x.first('1D').values
)

Considere los siguientes datos de acciones fundidas por minutos. No vemos los datos OHLC directamente:

melted_stock_data = pd.read_csv('data/melted_stock_data.csv', index_col='date', parse_dates=True)
melted_stock_data.head()

Podemos utilizar el método `ohlc()` después del remuestreo para recuperar las columnas OHLC:

melted_stock_data.resample('1D').ohlc()['price']

Alternativamente, podemos aumentar la muestra para aumentar la granularidad. Tenga en cuenta que esto introducirá valores `NaN`:

fb.resample('6H').asfreq().head()

Hay muchas maneras de manejar estos valores `NaN`. Podemos rellenar con `pad()`:

fb.resample('6H').pad().head()

Podemos especificar un valor concreto o un método con `fillna()`:

fb.resample('6H').fillna('nearest').head()

Podemos utilizar `asfreq()` y `assign()` para especificar la acción por columna:

fb.resample('6H').asfreq().assign(
    volume=lambda x: x.volume.fillna(0), # poner 0 cuando el mercado está cerrado
    close=lambda x: x.close.fillna(method='ffill'), # llevar adelante
    # tomar el precio de cierre si estos no están disponibles
    open=lambda x: x.open.combine_first(x.close),
    high=lambda x: x.high.combine_first(x.close),
    low=lambda x: x.low.combine_first(x.close)
).head()

## Merging
Vimos ejemplos de fusión en el [`1-querying_and_merging.ipynb`](./1-querying_and_merging.ipynb). Sin embargo, todas coincidían en función de las claves. Con las series temporales, es posible que sean tan granulares que nunca tengamos la misma hora para varias entradas. Trabajemos con algunos datos bursátiles en diferentes granularidades:

import sqlite3

with sqlite3.connect('data/stocks.db') as connection:
    fb_prices = pd.read_sql(
        'SELECT * FROM fb_prices', connection, 
        index_col='date', parse_dates=['date']
    )
    aapl_prices = pd.read_sql(
        'SELECT * FROM aapl_prices', connection, 
        index_col='date', parse_dates=['date']
    )

Los precios de Facebook son al minuto:

fb_prices.index.second.unique()

Sin embargo, los precios de Apple tienen información para el segundo:

aapl_prices.index.second.unique()

Podemos realizar una fusión *as of* para intentar alinearlas lo mejor posible. Especificamos cómo manejar el desajuste con los parámetros `direction` y `tolerance`. Vamos a rellenar con la "dirección" de "más cercano" y una "tolerancia" de 30 segundos. Esto colocará los datos de Apple con el minuto al que esté más cerca, así que 9:31:52 irá con 9:32 y 9:37:07 irá con 9:37. Como los tiempos están en el índice, pasamos `left_index` y `right_index`, como hicimos con `merge()` anteriormente en este capítulo:

pd.merge_asof(
    fb_prices, aapl_prices, 
    left_index=True, right_index=True, # las fechas están en el índice
    # fusionar con el minuto más cercano
    direction='nearest', tolerance=pd.Timedelta(30, unit='s')
).head()

Si no queremos perder la información de los segundos con los datos de Apple, podemos utilizar `pd.merge_ordered()` en su lugar, que intercalará los dos. Tenga en cuenta que se trata de una unión externa por defecto (parámetro `how`). La única pega es que tenemos que restablecer el índice para poder unirnos a él:

pd.merge_ordered(
    fb_prices.reset_index(), aapl_prices.reset_index()
).set_index('date').head()

Podemos pasar un `fill_method` para manejar valores `NaN`:

pd.merge_ordered(
    fb_prices.reset_index(), aapl_prices.reset_index(),
    fill_method='ffill'
).set_index('date').head()

Alternativamente, podemos utilizar `fillna()`.

<hr>

<div style="display: flex; justify-content: space-between; margin-bottom: 10px;">
    <div style="text-align: left;">
        <a href="./3-agregaciones.ipynb">
            <button>&#8592; Notebook Anterior</button>
        </a>
    </div>
    <div style="text-align: center;">
        <a href="../../solutions/ch_04/solutions.ipynb">
            <button>Soluciones</button>
        </a>
    </div>
    <div style="text-align: right;">
        <a href="../ch_05/1-introduccion_matplotlib.ipynb">
            <button>Capitulo 5 &#8594;</button>
        </a>
    </div>
</div>

<hr>
